Move everything into cogs, needs more testing before PR
